Cleaning up tests for readability.
