+++ src/IntegrationTests/MergeQuery/MergeQueryObjectTestsForRemappingBehavior.cs
@@ -50,14 +50,14 @@ namespace ivaldez.Sql.IntegrationTests.MergeQuery
 
             helper.DataService.Merge(request);
 
-            var databaseDtos = helper.DataService.GetAllSampleSurrogateKey().ToArray();
+            var sourceDtos = helper.DataService.GetAllSampleSurrogateKey().ToArray();
 
-            var firstDto = databaseDtos.First(x => x.TextValue == "JJ");
+            var firstDto = sourceDtos.First(x => x.TextValue == "JJ");
             firstDto.Pk.Should().BeGreaterThan(0);
             firstDto.IntValue.Should().Be(100);
             firstDto.DecimalValue.Should().Be(100.99m);
 
-            var secondDto = databaseDtos.First(x => x.TextValue == "ZZ");
+            var secondDto = sourceDtos.First(x => x.TextValue == "ZZ");
             secondDto.Pk.Should().BeGreaterThan(0);
             secondDto.IntValue.Should().Be(999);
             secondDto.DecimalValue.Should().Be(123.45m);
