+++ src/IntegrationTests/MergeQuery/MergeQueryObjectTestsForSourceTableTypes.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System.Collections.Generic;
+using System.Linq;
 using FluentAssertions;
 using ivaldez.Sql.IntegrationTests.Data;
 using ivaldez.Sql.SqlMergeQueryObject;
@@ -11,7 +12,7 @@ namespace ivaldez.Sql.IntegrationTests.MergeQuery
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
-        public void ShouldExecuteWithTempTable(bool useRealTempTable)
+        public void ShouldExecuteWithDefinedTempTableType(bool useRealTempTable)
         {
             var helper = new MergeQueryObjectTestHelper();
             helper.DataService.DropTable();
@@ -36,11 +37,24 @@ namespace ivaldez.Sql.IntegrationTests.MergeQuery
                 KeepPrimaryKeyInInsertStatement = false
             };
 
+            var infoLogger = new List<string>();
+            request.InfoLogger = info => infoLogger.Add(info);
+
             helper.DataService.Merge(request);
 
-            var databaseDtos = helper.DataService.GetAllSampleSurrogateKey().ToArray();
+            if (useRealTempTable == false)
+            {
+                infoLogger.Any(x => x.Contains("##MergeObjectTemp")).Should().BeTrue();
+            }
 
-            var firstDto = databaseDtos.First(x => x.TextValue == "JJ");
+            if (useRealTempTable == true)
+            {
+                infoLogger.Any(x => x.Contains("dbo.MergeObjectTemp")).Should().BeTrue();
+            }
+
+            var sourceDtos = helper.DataService.GetAllSampleSurrogateKey().ToArray();
+
+            var firstDto = sourceDtos.First(x => x.TextValue == "JJ");
             firstDto.IntValue.Should().Be(100);
             firstDto.DecimalValue.Should().Be(100.99m);
         }
